fix(bindings): remove -Wno-conversion, fix first-party warnings, mark eigenpy SYSTEM
Drop the blanket -Wno-conversion that hid genuine conversion bugs in our
binding code, then resolve what it exposed at the root cause:

- -Wsign-compare / -Wsign-conversion: use size_t loop counters in the
  container repr/str helpers (containers_export.hpp); explicit static_cast at
  boundaries (numerical_irreducible_decomposition.hpp)
- -Wunused-parameter: comment out unused PyClass&/object& params in visitors
  (configured_visitor, endgame_observers, tracker_observers, mpfr_export.cpp)
- -Wunused-local-typedef: drop the unused BRT alias (endgame_export.hpp)

eigenpy's own templates (numpy-allocator.hpp etc.) emit conversion/shorten
warnings when instantiated from our code, which marking the include dir SYSTEM
does not suppress (clang still reports warnings raised during template
instantiation requested from non-system code). Add
--system-header-prefix=eigenpy/ (clang-only) to treat eigenpy as a system
header, plus mark its include dirs SYSTEM. This silences ~120 third-party
eigenpy warnings without a blanket -Wno that would mask our own bugs.

One intentional -Wsign-conversion remains in PowVisitor<T,int>::__pow__: the
natural pow(a,b) is correct for every (T,S) here including negative integer
exponents (a**-2); forcing the exponent unsigned breaks that (caught by
mpfr_test::test_change_prec), and promoting to T fails to compile for complex T.

Full tree now builds with a single intentional warning. C++ suites pass;
pytest: 504 passed, 12 skipped, 1 pre-existing failure (a numpy 2.4 / Python
3.14 custom-dtype reduction SystemError, reproducible on a bare 2-element
array, unrelated to this change).

+++ core/include/bertini2/nag_datatypes/numerical_irreducible_decomposition.hpp
@@ -63,7 +63,7 @@ namespace bertini {
 				std::vector<int> codims;
 				for (const auto& w : finished_witness_sets_)
 					if (std::find(codims.begin(), codims.end(), w.Dimension()) == codims.end())
-						codims.push_back(w.Dimension());
+						codims.push_back(static_cast<int>(w.Dimension()));
 
 				return codims;
 			}
